Reduce SPXY memory consumption
Rather than storing 3 NxN matrices we now store 2 in memory and discard one as soon as we performed the sum.
This could be reduced to just 1 NxN matrix by computing DY iteratively - but the added complexity is probably not worth it since a Lazy version of the algorithm is available.